Change in the initialisation analysis to treat variables which get type 1/2 but are then used in a discrete time context. E.g., the following program is correct:
fun f (x, z) =>
  let y = last x + 1 in
  present z -> f(y+1) init 0

even if f gets type 0 -> 0 because, y gets type 1/2, which
means "initialised at discrete time instant only", and so
receive type 0 under the present handler. For that, the
type of y is subtyped into alpha^1/2, with 1/2 as its minimal
element such that alpha^1/2 < 0 is a valid subtyping.