Fix of a subtle bug in the module Static.
- the type environment was not set properly
after the beta-reduction, keeping static
(but useless) entries;
- as the reduce function was using the Inline
module, the renaming was simply wrong as it
was relying on the typing environment only.

- The current version of Reduce is a bit werd
as it does a bit of inlining. Normally, it should
not an simply introduce a new function entry when
the value of f in (f e) is an explicit lambda abstraction.

- I will change either to that solution or extend
the module Reduce so that it also does renaming.