Bug fix in the Markfunction module.
- The module seems to work fine now.
- This forces me to modify the causality
analysis, adding subtype for the result of
the application.

All my difficulties came from the fact that
the causality order is strict, that is,
a < b means that a is strictly before b.

It was sufficient for the first-order case 
but make things extremely complicated and counter
intuitive for the higher-order case. I will change
this. To keep the simplification of types through
the computation of IO, the simplest is to
add a type *a (or a*) for "strictly after a"
(or #a) and:

H, [x1 : #a1,..., xn: #an] |- x1 = e1 and ... and xn = en

H, [x1: a1,...,xn:an] |- e : t
---------------------------------------------------------
 H |- let rec x1 = e1 and ... and xn = en in e

H |- x: #t  H |- e : t
----------------------
H |- x = e