For the causality analysis, we impose an extra constraint for patterns.
In a pattern, e.g., (x1,..., xn, ...), the
causality type for variables must be
atomic. This is because we do not do complicated
distribution rules before static scheduling.

As an example, consider:

let f (x) = let x1, x2 = x in x1, x2

versus:

let g(x1, x2) = x1, x2

In the first one (f), x is considered to be
atomic so its causality type is a * a.

The causality signature is: val f : a * a -> a * a

In the second, x1 is atomic, with some
type a1; x2 is atomic with some type a2.

The causality signature is: val f : a * b -> a * b

We may remove this extra constraint in the
future, using the IO information to know
when and how a variable must be split to make
the program statically schedulable.