add module nodeS
it provides a way to turn a hybrid node into a discrete one like node but it also computes a matrix sens of partial derivatives of continuous state variables w.r.t. the inputs of the node being simulated. The hybrid node should be of type float array -C-> float array.