Some rewriting of the matcheq function. Still some trouble in the representation of tuples.
The old (and primary) solution which
consisted to have a big bag of values
with tuples that can have element being
bot or nil appear to be simpler with less
code to write; versus the new one with
a distinction of functions producing only
pvalues and nodes producing values.

I need to clean all this. The matcheq
function is still not correct and complete.