Start rewriting the manipulation of tuples.
- Should reconsider the definition of
values to include tuples; that would be
better/simpler.

- For the moment:

the expression (e1,..., en) is synchronising, that is, it is the smashed product.

- when defining a function:

node f (x1,...,xn) returns (o1,..., om)

f should apply to a tuple of length n, that is:

f(e1,...,en). In that case, the argument is
not a smashed product.

This is a bit ad-hoc. The reason not to have a general form of product (with no distinguished form of smashed product) is to avoid having to use a type argument to bound the number of iteration for fix-points; and simplify the administrative code when calling a strict functions. This is no gravé-dans-le-marbre.