Fixed an important bug in the initialization and causality analysis
Thanks ediff!
The code was completely wrong. When I copied the code from the
main branch, I introduced a bug.

The instance function was called when typing a variable in a pattern. This is wrong!

The (possible) reason is that the
instantiation function is also doing sub-typing, which is a bad idea.