Compile-time evaluation of constant expressions
Some progress.
- currently, the table of values
is stored into a file f.zlo as it was
expected for ZRun.
- it should be done differently and
deal with the case where a module
is opened without giving an implementation. All values in that
case cannot be constant values.
- rewrite/correct the definition of
[implementation] in module Const. All
constant values must be stored in the table but will do not generate any code.