Update reset.ml
- very simple rule which removes
reset eq every e when eq is
stateless.
- a distribution rule is implemented
but commented for the moment.
- the reason is that it needs a fusion
rule (like for clocked equations) but
I am not really sure this is better than
the current version where
reset eq every e is kept instead of a propagation to elementary stateful equations (init x = e and x = f(e) where f is a stateful function).