Extra field for assertions
Extension to deal with transparent
assertions.

- add free variables and
hidden state variables.

By default, assertions are treated in
a classical manner, that is, they
are compiled and executed with
the model to which they belong to.

For transparent assertions (only
for hybrid models), they are compiled
as independent nodes. We add extra
informations as for hybrid nodes.