Change in inout.ml
- start an important change in
the definition of machines in obc.

- the global parameter of a machine
is the first argument of the allocation
function.

machine p1..pn as self =
  initialization
    e (* can depend on (p1,...) *)
  memory
    ...
  instances
    fi : e1 p1;...
  methods
    step f1(x1) = ...
  ...
  assertion =
    [mach1;...;mach_k]
  end

The allocation function in OCaml
has the form:

let alloc (p1,...pn) = ... in { ... } in
...

The move to the new version is not
entirely made. For the moment, the
code is not correct for hybrid nodes,
in particular with transparent assertions.