Update - inout is now adapted to the new obc
- the parameters of a machines
are executed with the initialization code.
- the inout pass adds an
extra parameter cstate.
- every memory state of a hybrid
machine points to it with
a field name cstate_...;
- in a next version, we plan to
remove useless copies of
continuous state variables so that
the code points directly to the continuous state vector in the
cstate data-structure (yet, it is
not clear that there is a efficiency gain
to hope here; and it may compromise
having the size of the continuous
state vector to change dynamically.